use of long comes from f2c and is  32/64-bit inconsistent
git-svn-id: https://svn.r-project.org/R/trunk@64670 00db46b3-68df-0310-9c12-caf00c1e9a41

+++ src/nmath/bessel_j.c
@@ -1,6 +1,6 @@
 /*
  *  Mathlib : A C Library of Special Functions
- *  Copyright (C) 1998-2012 Ross Ihaka and the R Core team.
+ *  Copyright (C) 1998-2014 Ross Ihaka and the R Core team.
  *
  *  This program is free software; you can redistribute it and/or modify
  *  it under the terms of the GNU General Public License as published by
@@ -32,11 +32,11 @@
 
 #define min0(x, y) (((x) <= (y)) ? (x) : (y))
 
-static void J_bessel(double *x, double *alpha, long *nb, double *b, long *ncalc);
+static void J_bessel(double *x, double *alpha, int *nb, double *b, int *ncalc);
 
 double bessel_j(double x, double alpha)
 {
-    long nb, ncalc;
+    int nb, ncalc;
     double na, *bj;
 #ifndef MATHLIB_STANDALONE
     const void *vmax;
@@ -59,7 +59,7 @@ double bessel_j(double x, double alpha)
          * this may not be quite optimal (CPU and accuracy wise) */
         return (bessel_j(x, -alpha) * cospi(alpha) + ((alpha == na) ? 0 : bessel_y(x, -alpha) * sinpi(alpha)));
     }
-    nb = 1 + (long)na; /* nb-1 <= alpha < nb */
+    nb = 1 + (int)na; /* nb-1 <= alpha < nb */
     alpha -= (double)(nb - 1);
 #ifdef MATHLIB_STANDALONE
     bj = (double *)calloc(nb, sizeof(double));
@@ -91,7 +91,7 @@ double bessel_j(double x, double alpha)
    allocating one. */
 double bessel_j_ex(double x, double alpha, double *bj)
 {
-    long nb, ncalc;
+    int nb, ncalc;
     double na;
 
 #ifdef IEEE_754
@@ -112,7 +112,7 @@ double bessel_j_ex(double x, double alpha, double *bj)
         return (bessel_j_ex(x, -alpha, bj) * cospi(alpha) +
                 ((alpha == na) ? 0 : bessel_y_ex(x, -alpha, bj) * sinpi(alpha)));
     }
-    nb = 1 + (long)na; /* nb-1 <= alpha < nb */
+    nb = 1 + (int)na; /* nb-1 <= alpha < nb */
     alpha -= (double)(nb - 1);
     J_bessel(&x, &alpha, &nb, bj, &ncalc);
     if (ncalc != nb)
@@ -127,7 +127,7 @@ double bessel_j_ex(double x, double alpha, double *bj)
     return x;
 }
 
-static void J_bessel(double *x, double *alpha, long *nb, double *b, long *ncalc)
+static void J_bessel(double *x, double *alpha, int *nb, double *b, int *ncalc)
 {
     /*
      Calculates Bessel functions J_{n+alpha} (x)
@@ -242,7 +242,7 @@ static void J_bessel(double *x, double *alpha, long *nb, double *b, long *ncalc)
                                     6.2044840173323943936e23};
 
     /* Local variables */
-    long nend, intx, nbmx, i, j, k, l, m, n, nstart;
+    int nend, intx, nbmx, i, j, k, l, m, n, nstart;
 
     double nu, twonu, capp, capq, pold, vcos, test, vsin;
     double p, s, t, z, alpem, halfx, aa, bb, cc, psave, plast;
@@ -270,7 +270,7 @@ static void J_bessel(double *x, double *alpha, long *nb, double *b, long *ncalc)
                 b[i] = 0.; /*was ML_POSINF (really nonsense) */
             return;
         }
-        intx = (long)(*x);
+        intx = (int)(*x);
         /* Initialize result array to zero. */
         for (i = 1; i <= *nb; ++i)
             b[i] = 0.;
